Merge Release 1.2.0 changes into Main

+++ FixAntenna/NetCore/Common/Utils/ByteBuffer.cs
@@ -165,7 +165,8 @@ namespace Epam.FixAntenna.NetCore.Common.Utils
 				IncreaseBuffer(64);
 			}
 
-			_writer.Write(b);
+			_writer.Write((byte)b);
+
 			return this;
 		}
